global architecture
