+++ src/eolabtools/detection_orientation_culture/detection_orientation_culture/generate_slope_aspect.sh
