+++ src/eolabtools/detection_orientation_culture/environment.yml
