+++ src/eolabtools/night_osm_registration/AUTHORS.rst
