+++ src/eolabtools/sun_map_generation/src/SunMapGenerator.py
