--- src/sun_map_generation/listings/.ipynb_checkpoints/test1-checkpoint.lst
@@ -1,1 +1,0 @@
-/work/EOLAB/DATA/MNS/Paris/75-2021-0653-6866-LA93-0M50.tif
