--- src/sun_map_generation/listings/listing_test.lst
@@ -1,9 +1,0 @@
-/work/EOLAB/DATA/MNS/Paris/75-2021-0648-6861-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0648-6862-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0648-6863-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0649-6861-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0649-6862-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0649-6863-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0650-6861-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0650-6862-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0650-6863-LA93-0M50.tif
