--- src/sun_map_generation/listings/test1.lst
@@ -1,1 +1,0 @@
-/work/EOLAB/DATA/MNS/Paris/75-2021-0648-6862-LA93-0M50.tif
