--- src/sun_map_generation/listings/test4.lst
@@ -1,3 +1,0 @@
-/work/EOLAB/DATA/MNS/Paris/75-2021-0645-6863-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0646-6863-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0647-6863-LA93-0M50.tif
