--- src/sun_map_generation/listings/test5.lst
@@ -1,10 +1,0 @@
-/work/EOLAB/DATA/MNS/Paris/75-2021-0653-6866-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0651-6865-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0647-6861-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0647-6862-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0659-6860-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0651-6867-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0654-6860-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0649-6862-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0651-6863-LA93-0M50.tif
-/work/EOLAB/DATA/MNS/Paris/75-2021-0647-6864-LA93-0M50.tif
